python/requirements_yahooFinance.txt
PySocks==1.7.1
certifi==2019.9.11
chardet==3.0.4
functions-framework==3.*
protobuf==3.20.*
google-api-core==1.22.1
google-auth==1.20.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.4.1
google-cloud-pubsub==1.7.0
google-cloud-storage==1.31.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
idna==2.8
oauthlib==3.1.0
requests-oauthlib==1.3.0
requests>=2.22.0
six==1.13.0
yfinance==0.1.74
urllib3==1.25.7